formatted dates to show in simplified view instead of all
