+++ src/utils.js
@@ -1,5 +1,6 @@
 import { timeParse, bisector } from "d3";
 import { enqueueSnackbar } from 'notistack';
+import moment from 'moment';
 
 export const parseDate = timeParse("%d-%m-%Y");
 
@@ -72,4 +73,69 @@ export const findCAGRByNAV = (startNav, endNav, time) => {
   startNav = parseFloat(startNav);
   endNav = parseFloat(endNav);
   return parseFloat((((endNav / startNav) ** (1 / time) - 1) * 100).toFixed(2));
+};
+
+export const calculateRollingReturns = (navData, rollingPeriod = 3, totalRange = 10) => {
+  if (rollingPeriod > totalRange) {
+    return {
+      message: 'Rolling Period exceeded',
+      statusCode: 403,
+      data: {}
+    };
+  }
+
+  const today = moment();
+  const startDate = today.clone().subtract(12 * totalRange, 'months');
+  let [startDateNav, startDateIndex] = findNavByDate(navData, startDate.format('DD-MM-YYYY'));
+
+  if (!startDateNav) {
+    return {
+      message: `Scheme is started on ${navData[navData.length - 1]?.date}`,
+      statusCode: 403,
+      data: {}
+    };
+  }
+
+  const rollingEndDate = today.clone().subtract(12 * (totalRange - rollingPeriod), 'months');
+  const [rollingEndDateNav, rollingEndIndex] = findNavByDate(navData, rollingEndDate.format('DD-MM-YYYY'));
+
+  const rollingCAGRs = [];
+  const startDates = [];
+  for (let endIndex = rollingEndIndex; endIndex >= 0; endIndex--) {
+    const startNav = navData[startDateIndex]?.nav;
+    startDates.push(navData[startDateIndex]?.date);
+    startDateIndex -= 1;
+    const endNav = navData[endIndex]?.nav;
+    const cagr = findCAGRByNAV(startNav, endNav, rollingPeriod);
+    rollingCAGRs.push(cagr);
+  }
+
+  const [minCagr, maxCagr, meanCagr, medianCagr] = minMaxMeanMedian(rollingCAGRs);
+  const percentageDistribution = {
+    "<0%": 0, "0-8%": 0, "8-12%": 0, "12-15%": 0, "15-20%": 0, ">20%": 0
+  };
+
+  rollingCAGRs.forEach(cagr => {
+    if (cagr < 0) percentageDistribution['<0%']++;
+    else if (cagr <= 8) percentageDistribution['0-8%']++;
+    else if (cagr <= 12) percentageDistribution['8-12%']++;
+    else if (cagr <= 15) percentageDistribution['12-15%']++;
+    else if (cagr <= 20) percentageDistribution['15-20%']++;
+    else percentageDistribution['>20%']++;
+  });
+
+  const todayNav = navData[0]?.nav;
+  return {
+    [`${totalRange}Y CAGR`]: findCAGRByNAV(startDateNav, todayNav, totalRange),
+    avgCAGR: meanCagr,
+    minCAGR: minCagr,
+    maxCAGR: maxCagr,
+    medianCAGR: medianCagr,
+    percentageDistribution: percentageDistribution,
+    percentageDistributionPercent: Object.fromEntries(
+      Object.entries(percentageDistribution).map(([key, value]) => [key, parseFloat((value * 100 / rollingCAGRs.length).toFixed(2))])
+    ),
+    dates: startDates,
+    cagrs: rollingCAGRs
+  };
 };
